requirements.txt
flask
flask_table
flask-script
flask-bootstrap
flask-sqlalchemy
flask-wtf
wtforms_components
flask-login
Flask-Datepicker
